refactor feedback loop logic into it's own component
