+++ packages/component/src/Attachment/Text/private/MarkdownTextContent.tsx
@@ -26,6 +26,8 @@ import isAIGeneratedActivity from './isAIGeneratedActivity';
 import isBasedOnSoftwareSourceCode from './isBasedOnSoftwareSourceCode';
 import isHTMLButtonElement from './isHTMLButtonElement';
 import ActivityFeedback from '../../../Activity/ActivityFeedback';
+import hasFeedbackLoop from '../../../Activity/private/hasFeedbackLoop';
+import ActivityFeedbackLoop from '../../../Activity/private/ActivityFeedbackLoop';
 
 const { useLocalizer, useStyleOptions } = hooks;
 
@@ -246,9 +248,13 @@ const MarkdownTextContent = memo(({ activity, children, markdown }: Props) => {
         {activity.type === 'message' && activity.text && messageThing?.keywords?.includes('AllowCopy') ? (
           <ActivityCopyButton className="webchat__text-content__activity-copy-button" targetRef={contentRef} />
         ) : null}
-        {activity.type === 'message' && feedbackActionsPlacement === 'activity-actions' && (
-          <ActivityFeedback activity={activity} />
-        )}
+        {activity.type === 'message' &&
+          feedbackActionsPlacement === 'activity-actions' &&
+          (hasFeedbackLoop(activity) ? (
+            <ActivityFeedbackLoop activity={activity} />
+          ) : (
+            <ActivityFeedback activity={activity} />
+          ))}
       </div>
     </div>
   );
